Circuit Diagram excluding BlueTooth
Includes everything else: cap touch circuit, driver circuit, thermistor, batteries.

Right now has a resistor as a stand in for the hair and says "this is the hair". Planning to doctor the image in Adobe Illustrator to include an illustrated braid. Also going to outline the different circuits within the image.